[SPARK-31125][K8S] Terminating pods have a deletion timestamp but they are not yet dead
Change what we consider a deleted pod to not include "Terminating"

If we get a new snapshot while a pod is in the process of being cleaned up we shouldn't delete the executor until it is fully terminated.

No

This should be covered by the decommissioning tests in that they currently are flaky because we sometimes delete the executor instead of allowing it to decommission all the way.

I also ran this in a loop locally ~80 times with the only failures being the PV suite because of unrelated minikube mount issues.

Closes #27905 from holdenk/SPARK-31125-Processing-state-snapshots-incorrect.

Authored-by: Holden Karau <hkarau@apple.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 57d27e900f79e6c5699b9a23db236aae98e761ad)
Change-Id: Ic4af68f5930dee74a44a3de47aa5c6d3e1143e89