CDPD-9597. Enable spark to work with OpenCSVSerde hive tables after HIVE-22920.
Before HIVE-22920, hive metastore wasn't managing the metadata of OpenCSVSerde tables and hive was considering data type of each field in schema as string for such tables. As a rsult spark was finding a difference between its own schema (which uses the actual data type of the field)  and hive shema. In this situation spark will consider the hive schema to be correct.

After HIVE-22920, Hive will store the correct schema of OpenCSVSerde tables in metastore. Meaning that each field's data type of a OpenCSVSerde table will be stored in metastore as it is definied during table creation. So there will be no difference between shemas in hive and spark. However OpenCSVSerde considers all the column to be of the type of string. Spark needs to check the type of each field and do proper casting between string and the actual type of the field.

This change provides especial case handlings for OpenCSVSerde tables.
Testing has been done on actual cluster and spark/hive can read/write OpenCSVSerde tables created by hive/spark after this change.

Change-Id: Ic5b0014aaa3cdfa769f1a4d603594f4e7293a2c2