CDPD-13213. Migrate regression UT from SPARK-29999 to avoid empty partition issue in file stream sink
It turned out the root cause of CDPD-11996 was SPARK-26081 which is known to break file stream sink
in case when there's no row to write for any partition.

There's SPARK-29999 which fixed the issue - the change itself was reverted with SPARK-26081 due to
overhead, but the UT checking regression wasn't reverted.

This commit introduces the UT to avoid regression here as well.

Change-Id: I9f356bfe96aef466cc58d40c88fa8c9c61c022c6