CDPD-10247. Deps & Shading required for Raz integration.
Raz requires jersey1, which conflicts with spark's use of jersey2.  So we shade jersey1 and repackage the raz clients to reference our shaded jersey1.  Similarly for raz's use of jackson-jaxrs.  Finally, as raz is pulled in transitively via hadoop-cloud-storage, the assembly has to exclude the transitive dependency, and instead pull in our shaded version.

After these changes, the raz classes are in their original packages (so the original hadoop-cloud-storage can still reference them), but they contain references to shaded jersey1 & jackson classes.

Signed-off-by: Ramesh Mani <rmani@cloudera.com>

Change-Id: I6b061919ffa79bdda3f2f32c72fa87c295e70b4e