CDPD-14522. Exclude Apache Shiro from kafka-clients' dependencies.
We pin kafka-clients' version to a specific build, and that build uses
a version of Apache Shiro that has a vulnerability. The Kafka team has
upgraded the version of Apache Shiro that they use. However, since we
pin to a specific build of kafka-clients, we won't benefit from  that
change.

Change-Id: I2accc61f03473b81d6049077ca7bb292fa67ff3b