CDPD-15950. Add alternate config for a config renamed in Spark 3.
spark.eventLog.logStageExecutorMetrics.enabled was renamed to
spark.eventLog.logStageExecutorMetrics just before Spark 3.0
was released. This config is new to 3.0. However, we have already
backported it to Spark 2.4 in some CDH 6 and 7 versions (under
the old name).
To ease migration, we add the old name as an alternate config
for the new name.

Change-Id: I2c24d5ae1c51acb4bcc193c2edb1fca134c66744