CDPD-17235. Deps & Shading required for raz-s3 integration for Spark.
Exclude raz-hook-s3 and raz-s3-lib modules from assembly/pom.xml and hadoop-cloud/pom.xml and include those dependencies in shaded-raz/pom.xml.
This is needed because raz requires jersey1, which conflicts with spark's use of jersey2.  So we shade jersey1 and repackage the raz clients to reference our shaded jersey1.

Change-Id: I1584c3fcd1fd708fefd351d2a6fa898abb0673a3