CDPD-18142. Fix subquery canonicalization errors after whole plan reuse
This is a follow-up ticket to CDPD-13679 and contains the latest commits from the preview of SPARK-29375, SPARK-28940 and SPARK-32041 (https://github.com/apache/spark/pull/28885).

This change modifies the canonicalization code of `ScalarSubquery` to avoid NPE during canonicalization when some of the transient fields (like `sqlContext`) is not available. These issues came up with the upstream PR (Spark 3.1) where UTs cover more scenarios now than at the time of CDPD-13679 was created.

Change-Id: I6b336ead4d760d4ab1ecef06dc1ecc370a7dda89