[CDPD-25453][SPARK-32978][SQL] Make sure the number of dynamic part metric is correct
The purpose of this pr is to resolve SPARK-32978.

The main reason of bad case describe in SPARK-32978 is the `BasicWriteTaskStatsTracker` directly reports the new added partition number of each task, which makes it impossible to remove duplicate data in driver side.

The main of this pr is change to report partitionValues to driver and remove duplicate data at driver side to make sure the number of dynamic part metric is correct.

The the number of dynamic part metric we display on the UI should be correct.

No

Add a new test case refer to described in SPARK-32978

Closes #30026 from LuciferYang/SPARK-32978.

Change-Id: Ife5932241612850987a931322fd3e38989ec7eba
Authored-by: yangjie01 <yangjie01@baidu.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>