[SPARK-33716][K8S] Fix potential race condition during pod termination
Check that the pod state is not pending or running even if there is a deletion timestamp.

This can occur when the pod state and deletion timestamp are not updated by etcd in sync & we get a pod snapshot during an inconsistent view.

No

Manual testing with local version of Minikube on an overloaded computer that caused out of sync updates.

Closes #30693 from holdenk/SPARK-33716-decommissioning-race-condition-during-pod-snapshot.

Authored-by: Holden Karau <hkarau@apple.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit bf2c88ccaebd8e27d9fc27c55c9955129541d3e1)
Change-Id: I7dc25af16b936dc2dedb2c3baf7069ff7965c8f0