[CDPD-24643][SPARK-33793][TESTS] Introduce withExecutor to ensure proper cleanup in tests
This PR introduces a helper method `withExecutor` that handles the creation of an Executor object and ensures that it is always stopped in a finally block. The tests in ExecutorSuite have been refactored to use this method.

Recently an issue was discovered that leaked Executors (which are not explicitly stopped after a test) can cause other tests to fail due to the JVM being killed after 10 min. It is therefore crucial that tests always stop the Executor. By introducing this helper method, a simple pattern is established that can be easily adopted in new tests, which reduces the risk of regressions.

No

Run the ExecutorSuite locally.

Closes #30783 from sander-goos/SPARK-33793-close-executors.

Authored-by: Sander Goos <sander.goos@databricks.com>
Signed-off-by: HyukjinKwon <gurwls223@apache.org>
(cherry picked from commit ddff94fd32f85072cbc5c752c337f3b89ae00bed)
Signed-off-by: HyukjinKwon <gurwls223@apache.org>
(cherry picked from commit 50e72dcda04db2d6741465ce7864ecaefa5f55e6)
Change-Id: I1f3bd2b3b49bb33083e9a13ab596fff938be2bda