[CDPD-11254] add ubi8 image base to spark 3
The refactoring of the scripts makes easier to introduce the build
of other OS based container images for Spark3.

A new build.sh script was introduced which builds
the Spark3 container images for one specific base OS.

The ubi8 and slim images are build in the build context
created and used by the docker-image-tool.sh.
The ubi8 and slim Dockerfiles and entrypoint.sh script
is copied there because the docker build commands
are executed in that directory. Every file
which is copied into image the should be copied from there.

The default slim image name was changed (extended with `-slim`) to be
more explicit and because the `bin/docker-image-tool.sh` script builds the
images with a name not including the OS.
It can be confusing when both slim and ubi8 images are built the name for the
ubi8 image will be `spark` and `spark-py` too but previously the slim images
were pushed with these names.  To not to push the wrong image by accident and
to create less confusion within the names the docker rm commands remove the
default image tags and only the image tags/names including the OS are kept.

Change-Id: Ie11f47ea42e4fcab3a8ef9550246ffe95f061572