[DEX-3021] smaller dex runtime image
Modifying the owner and group adds additional layers to the final docker image.
With this change every spark related file will be owned by the right
DEX user id 1345 so it is not needed to modify the the ownership of the files
when the dex specific images are built.
The file ownership change adds and extra layer during the image build
which increases the image size with ~500MB.
Without the extra ownership changed layer the final image size can be smaller
when deployed.

COPY --chown is only supported on Dockerfiles used to build Linux containers.

Change-Id: I432bbc4257b384b9deba3f97d893dcaa53c1099a