CDPD-22638. Correct backport of SPARK-29287.
The CoarseGrainedSchedulerBackend keeps a map, executorDataMap, of
executor id to ExecutorData (bookkeeping data that includes freeCores
(cores available for tasks) and totalCores (total number of cores the
executor has)).
Before SPARK-29287, when an executor sends a RegisterExecutor message
to the driver, the executor data is put in the executorDataMap, and
the driver can make offers with this executor.
With SPARK-29287, the driver does not make offers on receiving the
RegisterExecutor message. A new LaunchedExecutor message is introduced,
which the executor backend sends to the driver when the executor is
actually constructed, and then the driver makes offers with the
executor.
In order to prevent the executor from being available for offers
before LaunchedExecutor is received, the ExecutorData created for the
executor on RegisterExecutor should have freeCores set to 0, and
when LaunchedExecutor is received, the freeCores is set to totalCores,
and all the executor cores are then available for tasks.

When SPARK-29287 was backported along with other commits, a change in
one of the lines was incorrectly resolved in the conflict resolution.
This corrects the line.

Change-Id: I8a699906444f757b2e74149dee9bfd832ac82062