CDPD-24399. Improve batch size for initial pod requests when gang scheduling is available
To take advantage of place-holder pods in pre-allocated nodes (Yunikorn's "gang scheduling"),
we want ExecutorPodsAllocator to use a different batch size for the initial batch of requests
than it uses for subsequent batches. Ideally, the initial batch size will match the number of
place-holder pods. The number of placeholder pods will be the max of min-executors,
initial-executors, and num-Executors.

Change-Id: Ia342a1508885cfa4ea4e6e298e3599ba183bdf0a