[CDPD-25568][SPARK-30766] Fix the timestamp truncation to the HOUR and DAY levels
This is a backport of SPARK-30766 from Spark3. Cherry-picking the commit
is not an option here since it depends on some commits that modify
DateTimeUtils. Cherry-picking these dependencies will lead to a major
refactor of DateTimeUtils which might further introduce a regression.
Thus only handpicking the pieces of the Spark3 commit which are absolutely
necessary.

Change-Id: Iced59b27f18f47a7c50cc73f95d4d41f75be0a74