CDPD-26245. Run all unit tests even if a test fails.
Currently, if a unit test fails during pre-commit, maven skips over the
rest of the modules which list the failed module as a dependency.
E.g.
```
[INFO] Spark Project Unsafe ............................... SUCCESS [11.744 s]
[INFO] Spark Project Launcher ............................. SUCCESS [02:28 min]
[INFO] Spark Project Core ................................. FAILURE [21:34 min]
[INFO] Spark Project ML Local Library ..................... SUCCESS [32.438 s]
[INFO] Spark Project GraphX ............................... SKIPPED
[INFO] Spark Project Streaming ............................ SKIPPED
[INFO] Spark Project Catalyst ............................. SKIPPED
[INFO] Spark Project SQL .................................. SKIPPED
```

Sometimes developers will merge a Gerrit when the cause of the unit test
failure is known. However, they may not realize that not all unit tests
have actually run.
At the time of creating this PR, CDPD-25958 (Flatbuffers) and CDPD-26248
(HiveExternalCatalogSuite) are causing UT failures. Flatbufffer failure
occurs before HiveExternalCatalogSuite, if a developer doesn't know
about the latter issue, they can flag the first failure as a known issue
and merge the change without being aware of the fact that there is a
second issue at hand.

This change will run all the UTs in all modules ignoring the failure.
Once all tests are executed a python script validates if there are any
test errors or failures.

Change-Id: Ib87cbd1d41efe368f826052936e658c5638731f4