CDPD-27357. Filter all TS in filterInvalidTimestamps instead of just pre-gregorian TS
We generate Random timestamps for various UTs. Some of these
UTs run against all timezones, and due to this change in TZ not all of
the randomly generated timestamps are valid in all timezones.

In order to filter such TS we use
`DateTimeTestUtils.filterInvalidTimestamps`. This function just filters
pre-gregorian TS as an optimization, but we should enable the filtering
for all TS for complete correctness.

Change-Id: I6d9858b4cb338600cd273f105432e7fd3a77da6d