[SPARK-36129][BUILD] Upgrade commons-compress to 1.21 to deal with CVEs
This PR upgrades `commons-compress` from `1.20` to `1.21` to deal with CVEs.

Some CVEs which affect `commons-compress 1.20` are reported and fixed in `1.21`.
https://commons.apache.org/proper/commons-compress/security-reports.html

* CVE-2021-35515
* CVE-2021-35516
* CVE-2021-35517
* CVE-2021-36090

The severities are reported as low for all the CVEs but it would be better to deal with them just in case.

No.

CI.

Closes #33333 from sarutak/upgrade-commons-compress-1.21.

Cloudera ID: CDPD-28421

Authored-by: Kousuke Saruta <sarutak@oss.nttdata.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit fd06cc211d7d1579067ad717da9976aabd71b70d)
Change-Id: I779c323ead21eb0fa8324cadd12a1bdc0cc1302b