[DEX-5232] Dynamic allocation with RSS.
Introduce an option in Spark conf "spark.shuffle.registration.enabled" to explicitly declare if the executors
should register to the local external shuffle services. By default, it is set to true; when a remote shuffle
service is used, set this value to false in order to by pass the registration.

Upstream JIRA: SPARK-37394

Change-Id: I25b29d561a15cd8265656ada6584aef040edd12a